Django/requirements.txt
django==3.1.3

django-cors-headers==3.5.0

djangorestframework==3.12.2

mysqlclient==2.0.3

djangochannelsrestframework==0.3.0